Extract ask/quick helpers into focused modules
Slim src/app/api/ask/quick/route.ts (1316 -> 754 lines) by moving four
self-contained concerns into dedicated modules. Pure mechanical
extraction, no behavior change:

- services/org-canvas-conversation.ts: conversation-row lifecycle +
  prompt-cache helpers (resolveOrgConversationRowId,
  persistCanvasUserMessage, loadOrgCanvasPromptCache, hasConcepts,
  persistOrgCanvasPromptCache).
- lib/ai/resolveMessageImages.ts: pre-turn image-attachment URL
  resolution (relative presigned path -> absolute signed S3 URL).
- services/canvas-turn-enrichments.ts: post-turn follow-up questions +
  provenance Pusher broadcasts.
- lib/proposals/runProposalIntent.ts: synthetic Approve/Reject SSE
  stream.

These are now shareable by the upcoming non-streaming (agent-as-tool)
canvas turn path.
